feat(web): move unassigned docs from Overview into its own Documents tab
